fixed #1802 loading large keywords dictionary
git-svn-id: svn://svn.sphinxsearch.com/sphinx/branches/rel21@4601 406a0c4d-033a-0410-8de8-e80135713968

+++ src/sphinxint.h
@@ -1790,7 +1790,7 @@ struct InfixBlock_t
 		const char *	m_sInfix;
 		DWORD			m_iInfixOffset;
 	};
-	int			m_iOffset;
+	DWORD				m_iOffset;
 };
 
 
@@ -1802,7 +1802,7 @@ public:
 	virtual			~ISphInfixBuilder() {}
 	virtual void	AddWord ( const BYTE * pWord, int iWordLength, int iCheckpoint, bool bHasMorphology ) = 0;
 	virtual void	SaveEntries ( CSphWriter & wrDict ) = 0;
-	virtual int		SaveEntryBlocks ( CSphWriter & wrDict ) = 0;
+	virtual int64_t	SaveEntryBlocks ( CSphWriter & wrDict ) = 0;
 	virtual int		GetBlocksWordsSize () const = 0;
 };
